Renamed "CollisionShape2D" across the board to "Hitbox". Changed the Player scene to make it inheritance-friendly. Made the player sprite proportional again.
In case multiple CollisionShape2Ds have to exist on something, renamed all CollisionShape2Ds to "Hitbox" for clarity purposes. Centered the hitbox and the sprite of Player on (0,0) to make the scene inherit cleanly. The sprite for the player, for some reason, wasn't correctly proportioned, so I fixed it. Added a fire point position from which bullets can spawn from in preparation for implementing shooting.