set up some infrastructure for interactables
The player can now interact with things (default key: space). Interactable objects need the "interactable" tag as well as the functions interact(), toggleOff(), and toggleOn(). interact() gets called when the player interacts. The player can only interact with one interactable at a time, and so the interactable that is toggled on is the one that will be interacted with when the player interacts. Interactables also need to be a physics body.