Began work on AllyBase and added player follow to allies
A bunch of systems have been changed to do this, some of which amounted to things and others of which did not. The player has a list of allies and the function to spawn the ally is in the player's code itself. AllyBase is a base for specific ally types to inherit from. At current, it just follows the player around. The dummy interactables in the PlayerTester scene now spawn allies when interacted with.